requirements.txt
opencv-python==4.9.0.80
moviepy==1.0.3
Pillow==9.5.0
numpy==1.26.4
selenium
cairosvg
gTTS
ffmpeg